video: add yuv float formats
Adding all these so I can use them for obscure processing purposes,
maybe. There isn't really a reason why they should exist. On the other
hand, they're just labels for formats that can be handled in a generic
way, and this commit adds support for them in the zimg wrapper and
vo_gpu just by making the formats exist. (Well, vo_gpu had to be fixed
in the previous commit.)